pyproject.toml
[project]
name = "mcp-prometheus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
